# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: NOOBDY/questions_server/handlers
  Question:
    description: Question defines the structure for an API question
    properties:
      answer:
        type: string
        x-go-name: Answer
      option_a:
        type: string
        x-go-name: OptionA
      option_b:
        type: string
        x-go-name: OptionB
      option_c:
        type: string
        x-go-name: OptionC
      option_d:
        type: string
        x-go-name: OptionD
      question:
        type: string
        x-go-name: Question
    type: object
    x-go-package: NOOBDY/questions_server/data
info:
  description: Documentation for Question API
  title: Question API
  version: 1.0.0
paths:
  /questions:
    get:
      description: Returns a list of questions from the database
      operationId: listQuestions
      responses:
        "200":
          $ref: '#/responses/questionsResponse'
      tags:
      - questions
  /questions/{id}:
    get:
      description: Returns a single question from the database
      operationId: listQuestion
      parameters:
      - description: The id of the product for which the operation relates
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/questionResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - questions
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  questionResponse:
    description: Data structure representing a single question
    schema:
      $ref: '#/definitions/Question'
  questionsResponse:
    description: A list of questions
    schema:
      items:
        $ref: '#/definitions/Question'
      type: array
schemes:
- http
swagger: "2.0"